Reorganize files
After removing LRASolver, we can move the rest of the files from lrasolver folder to lasolver.
